refactor: improve tests
