+++ Cargo.toml
@@ -16,6 +16,7 @@ log = { version = "0.4.17" }
 serde = { version = "1.0.152", features = ["derive"] }
 miette = { version = "7.2.0", features = ["fancy"] }
 insta = { version = "1.38.0", features = ["filters", "glob"] }
+thiserror = "1.0.48"
 
 [profile.dev.package.insta]
 opt-level = 3
