+++ typechecker/Cargo.toml
@@ -6,15 +6,14 @@ edition = "2021"
 [dependencies]
 enderpy_python_parser = { path = "../parser", version = "0.1.0" }
 config = "0.13.3"
-serde = { version = "1.0.164", features = ["derive"] }
+serde.workspace = true
 miette.workspace = true
-thiserror = "1.0.48"
+thiserror.workspace = true
 log.workspace = true
-insta.workspace = true
 env_logger = "0.10.0"
 tempfile = "3.8.0"
-elsa = "1.10.0"
 is-macro = "0.3.5"
 bitflags = "2.4.2"
 
 [dev-dependencies]
+insta.workspace = true
